pyproject.toml
[project]
name = "expar-arflow-integration"
version = "0.1.0"
description = ""
authors = [
    {name = "Yiqin Zhao",email = "yiqinzhao@outlook.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">3.10,<3.13"
dependencies = [
    "arflow (>=0.3.0,<0.4.0)",
    "websockets (>=15.0,<16.0)",
    "black (>=25.1.0,<26.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
